Rename `DictTable` to `DictValue`
