+++ src/test/kotlin/sigma/syntax/expressions/OrderedTupleConstructorTermTests.kt
@@ -1,16 +1,9 @@
 package sigma.syntax.expressions
 
-import sigma.evaluation.scope.FixedScope
-import sigma.evaluation.scope.Scope
-import sigma.evaluation.values.BoolValue
-import sigma.evaluation.values.IntValue
 import sigma.evaluation.values.Symbol
-import sigma.evaluation.values.tables.ArrayTable
-import sigma.evaluation.values.tables.DictTable
 import sigma.syntax.SourceLocation
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlin.test.assertIs
 
 class OrderedTupleConstructorTermTests {
     object ParsingTests {
